Ignoring keyfiles and setting proper permissions
There is no need to have the keyfiles shared back to the repo by default since each user should have their own.

This change is being pushed by itself so it will be automatically picked up for the creation of the new filesystem images which include git so it will not need to be pulled every time.