Initialization+Load/Save Config Scripts & Jayanthi's Changes
Creation of scripts to handle initializing config files as well as the ability to store and retrieve them when new versions are pulled from the server.  This paves the way for maintaining a single branch rather than one for V2X/RSU and another for Modem.
Config files were added to .gitignore and are set --assume-unchanged each time the initialization or load scripts are run to minimize chances of them getting accidentaly pushed to the server.

Jayanthi's changes to RSU and V2X files implement a message counter among other improvements that help determine if packets are being dropped over DSRC.