Updates to support manual and OTA upgrade compatibility
Added proper exit codes to load and save config scripts.
Modified ota_upgrade python script such that it can now be run directly by the user to either manually install a new firmware version or force the device to poll the AWS server to check for a new version.  The script now uses the load and save shell scripts in the /scripts directory to backup and restore the user's configuration rather than handling it through the xc-upgrade shell script.  Speaking of xc-upgrade, since much of the tasks being performed by this script have been moved to the python script which calls it, and some new directories have been added (/scripts, /etc) new versions of this file will be used starting with the 2.1.6 firmware release.  These changes also fix the issue where every new firmware version overwrote the one in the factory directory.