Allow openxc_init to function from any directory as well as having it set permissions on key files
Fixed one of the paths in openxc_init to be absolute rather than relative so the program can be run from anywhere.  Also added a section to set permissions on all .pem files in the repo to 700 so scp will function properly.
Adding 2.1.5 tarball files to the current and factory backup directories so factory resets and software rollbacks during upgrade will work proerly as soon as the repo is cloned.