Fix for packet dropping issue
When packets were sent from the RSU parking demo null characters were included between the packets, presumably to indicate that they were a unit.  During the filtering and cleaning up of these packets after they were received by the V2X a problem could occur.  This only occurs if the timing is just right such that the message broadcast by the V2X ends up in its input buffer before the RSU packet. Under these conditions the filtering functions interpret the V2X packet as part of the RSU transmission, locate the V2X hostname in the transmission and discard it.  To the user this looks like the packet from the RSU never arrived when in reality it was just lost in translation.

This update includes changes to the parking garage class to remove the null characters between packets as well as eliminate the first empty packet.  On the V2X side, the python strip() method is called on each transmission to remove any hanging null or other undesired characters to make this issue less likely to reappear in the future.