Compliance with OpenXC Message Format
Whitespace changes in xc_vi.py so script will run without Python throwing errors
Added a few lines to each of the functions that handle sending data over DSRC, Bluetooth, and USB to ensure that all packets have trailing null characters.  This is in accordance with the message format specification and prevents errors in the mobile app or other endpoints for data generate by the accessory.
Modified how null characters are injected into the message stream and handled when they are encountered so that RSU data appears correctly in enabler (prevents "double message" errors)