Scripts for fake-hwclock Fully Under Version Control
Moved scripts and settings files for fake-hwclock to a central, version controlled location so that devices can get the functionality with a simple git pull rather than a full install.  Devices which will use this functionality will need symbolic links added from where the system is expecting to find these files to their new homes in the GIT repo.